+++ organizations/contentstrategies/client.go
@@ -6,11 +6,11 @@ import (
 	context "context"
 	os "os"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v21"
-	core "github.com/KardFinancial/kard-go-sdk/v21/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v21/internal"
-	option "github.com/KardFinancial/kard-go-sdk/v21/option"
-	organizations "github.com/KardFinancial/kard-go-sdk/v21/organizations"
+	kard "github.com/KardFinancial/kard-go-sdk/v22"
+	core "github.com/KardFinancial/kard-go-sdk/v22/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v22/internal"
+	option "github.com/KardFinancial/kard-go-sdk/v22/option"
+	organizations "github.com/KardFinancial/kard-go-sdk/v22/organizations"
 )
 
 type Client struct {